Update 2021-06-19-installing-and-running-wings-xi-on-linux.md
Changed the directions from using sed to just telling the user to use a text editor since it's more robust and the installer has changed some of the config stuff now that prevents the command from working reliabliy. 

I've opted not to change the command again as it could break in another installer and I would rather not deal with it again. :)